revert to use float

+++ data/api.py
@@ -50,7 +50,7 @@ class API:
 
 
 
-    def get_price_for_tickers(self, tickers: List[str], day: date) -> dict[str, Decimal]:
+    def get_price_for_tickers(self, tickers: List[str], day: date) -> dict[str, float]:
         """
         Args:
             tickers: list of ticker symbols to fetch the price for.
@@ -59,7 +59,7 @@ class API:
         Returns:
             The price table indexed by ticker symbol.
         """
-        table: dict[str, Decimal] = { }
+        table: dict[str, float] = { }
         conn = connect_to_database(self._host)
         stocks: List[Stock] = get_stocks(conn)
 
@@ -70,7 +70,7 @@ class API:
 
 
 
-    def get_price_history_for_tickers(self, tickers: List[str], start_date: date, end_date: date) -> dict[date, dict[str, Decimal]]:
+    def get_price_history_for_tickers(self, tickers: List[str], start_date: date, end_date: date) -> dict[date, dict[str, float]]:
         """
         Args:
             tickers: list of ticker symbols to fetch the price_history for
@@ -80,7 +80,7 @@ class API:
         Returns:
             the price history table indexed by date, then by ticker symbol
         """
-        matrix: dict[date, dict[str, Decimal]] = defaultdict(dict)
+        matrix: dict[date, dict[str, float]] = defaultdict(dict)
         conn = connect_to_database(self._host)
         stocks = get_stocks(conn)
 
@@ -91,7 +91,7 @@ class API:
 
 
 
-    def store_portfolio_results(self, portfolio: Portfolio, signals: dict[SignalBase, float], yearly_return: Decimal):
+    def store_portfolio_results(self, portfolio: Portfolio, signals: dict[SignalBase, float], yearly_return: float):
         """
         Caches the provided portfolio and its return to the database.
         Args:
