+++ data/stock_price.py
@@ -1,5 +1,4 @@
 ﻿from datetime import date, timedelta
-from decimal import Decimal
 from mysql.connector.abstracts import MySQLConnectionAbstract
 from data.stocks import Stock
 import pandas as pd
@@ -7,7 +6,7 @@ import pandas as pd
 
 def fill_stocks_price_history_matrix(
         conn: MySQLConnectionAbstract,
-        matrix: dict[date, dict[str, Decimal]],
+        matrix: dict[date, dict[str, float]],
         first_day: date,
         last_day: date,
         stocks: list[Stock]):
@@ -64,7 +63,7 @@ def fill_stocks_price_history_matrix(
 
 def fill_stocks_price_table(
         conn: MySQLConnectionAbstract,
-        table: dict[str, Decimal],
+        table: dict[str, float],
         day: date,
         stocks: list[Stock]):
     """
@@ -100,10 +99,10 @@ def fill_stocks_price_table(
         cursor.execute(sql, sql_params)
         for ticker, price in cursor.fetchall():
             if price is not None:
-                table[ticker] = Decimal(price)
+                table[ticker] = float(price)
 
 
-def get_stock_price(conn: MySQLConnectionAbstract, stock_id: str, when: date | str) -> Decimal or None:
+def get_stock_price(conn: MySQLConnectionAbstract, stock_id: str, when: date | str) -> float or None:
     """
     Fetches the closing price of a stock for a specific date.
